# docs/swagger.yaml
basePath: /
definitions:
  api.ErrResponse:
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
  model.Message:
    properties:
      fileUrl:
        type: string
      id:
        type: integer
      message:
        type: string
      phone:
        type: string
      sentAt:
        type: string
      status:
        type: string
      type:
        type: string
      userID:
        type: integer
    type: object
  model.MessageList:
    properties:
      items:
        items:
          $ref: '#/definitions/model.Message'
        type: array
      totalItemsCount:
        type: integer
    type: object
...
  model.StatusResponse:
    properties:
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for app
  title: lct2024 API
  version: "1.0"
paths:
  /message:
    get:
      description: Message List
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MessageList'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Get Message List
      tags:
      - Message
    post:
      description: Save Message
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.StatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrResponse'
      summary: Save Message
      tags:
      - Message
...
swagger: "2.0"